INSTALL.md: remove `plyvel` from `setup.py`, it fails on Python 3.9
As `plyvel` is included in `setup.py` it is downloaded
and installed by `pip`. This normally compiles `plyvel`
from source, which is the reason `libleveldb-dev`
(C++ headers) is required. Testing the compilation
with Python 3.9 results in errors, however.

We do not need to compile anything if we can just
install the prepackaged library for by Linux.
In Ubuntu this is `python3-plyvel`,
in Arch it is `python-plyvel`.

The original `setup.py` requests `plyvel` 1.0.5,
the packages provide a newer version,
1.2 (Ubuntu) and 1.3 (Arch).