setup.py: use `plyvel==1.2.0` instead of `1.0.5`
Compilation of `plyvel` 1.0.5 (July 17, 2018)
proceeds when using Python 3.7 and 3.8, but fails with 3.9.

If using `plyvel` 1.2 or 1.3, no compilation is necessary,
as the packages are downloaded from `pip` without issue.

Version 1.2.0 (January 22, 2020) is available in Ubuntu 20.04,
and 1.3.0 (October 10, 2020) is available in Arch,
so these are good versions to use as reference.