setup.py: use `aiohttp==3.6.2` instead of `aiohttp==3.5.4`
As reported in issue #2769, the `lbrynet` daemon doesn't respond
correctly when it is compiled against Python 3.8+, only with
Python 3.7.

Instead of using `aiohttp==3.5.4` which was released on
January 12, 2019, we use `aiohttp=3.6.2` which was released
October 9, 2019, and is available in Ubuntu 20.04.

By using this newer version, the RCP server seems to work
correctly both with Python 3.8 and 3.9.

We could also try a newer version, like `3.7.4` as
it was launched on February 25, 2021, and is available
in Arch.