file_manager: trap the exception if the claim is not a stream
If `claim` is not a stream, the line `claim.stream.source.bt_infohash`
will raise a `ValueError` exception because `claim.stream`
does not return a valid object.

This will happen if we try to download a claim (`lbrynet get`)
that is in fact a `collection` (list of streams).

Since at the moment we don't trap this exception, it prints
the entire trace of the exception.
```
Traceback (most recent call last):
  File "/opt/git/lbry-sdk/lbry/file/file_manager.py", line 117, in download_from_uri
    if claim.stream.source.bt_infohash:
  File "/opt/git/lbry-sdk/lbry/schema/claim.py", line 55, in stream
    return Stream(self)
  File "/opt/git/lbry-sdk/lbry/schema/claim.py", line 108, in __init__
    self.message = self.claim.get_message(self.claim_type)
  File "/opt/git/lbry-sdk/lbry/schema/claim.py", line 46, in get_message
    raise ValueError(f'Claim is not a {type_name}.')
```

Therefore we trap this `ValueError` to avoid showing the entire trace.
We don't do anything with the exception, we just raise it again
so that it is handled by the calling code `jsonrpc_get`.

Then it will show an appropriate message.
```
WARNING  lbry.extras.daemon.daemon:1106:
Error downloading lbry://@something#e/My-New-List#f: Claim is not a stream.
```